refactor: 즐겨찾기 유저별로 구분되도록 수정
