+++ src/containers/group/groupId/tasklist/DateNavigate.tsx
@@ -1,6 +1,7 @@
 import { IconCalendar } from '@/assets/IconList';
 import Calendar from '@/components/calendar/Calendar';
 import useQueryParameter from '@/hooks/useQueryParameter';
+import useUserStore from '@/stores/userStore';
 import getMonthDay from '@/utils/getMonthDay';
 import { useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
@@ -13,6 +14,7 @@ function DateNavigate() {
   const { groupId, taskListId } = router.query;
   // const { groupId, taskListId } = useQueryParameter();
   const queryClient = useQueryClient();
+  const { user } = useUserStore();
 
   return (
     <div className='flex items-center'>
@@ -43,6 +45,7 @@ function DateNavigate() {
                 queryClient,
                 Number(groupId),
                 Number(taskListId),
+                user?.id!,
               )
             }
           >
@@ -69,6 +72,7 @@ function DateNavigate() {
                 queryClient,
                 Number(groupId),
                 Number(taskListId),
+                user?.id!,
               )
             }
           >
