+++ src/containers/group/groupId/tasklist/ListPriority/useTaskPriority.ts
@@ -1,11 +1,14 @@
 import { getTaskPriority } from '@/services/TaskPriority.API';
+import useUserStore from '@/stores/userStore';
 import getMonthDay from '@/utils/getMonthDay';
 import { useQuery } from '@tanstack/react-query';
 
 function useTaskPriority(groupId: number, taskListId: number, date: string) {
+  const { user } = useUserStore();
+
   const { data, isLoading, refetch, isError, error } = useQuery({
-    queryKey: ['getPriorityTasks', groupId, taskListId, date],
-    queryFn: () => getTaskPriority(groupId, taskListId, date),
+    queryKey: ['getPriorityTasks', groupId, taskListId, date, user?.id!],
+    queryFn: () => getTaskPriority(groupId, taskListId, date, user?.id!),
     enabled: !Number.isNaN(groupId) && !Number.isNaN(taskListId),
   });
