+++ src/containers/group/groupId/useDateStore.ts
@@ -10,6 +10,7 @@ interface DateState {
     queryClient: QueryClient,
     groupId: number,
     taskListId: number,
+    userId: number,
   ) => void;
 }
 
@@ -21,6 +22,7 @@ const useDateStore = create<DateState>((set) => ({
     queryClient: QueryClient,
     groupId: number,
     taskListId: number,
+    userId: number,
   ) =>
     set((state) => {
       const newDate = new Date(state.pickDate);
@@ -39,6 +41,7 @@ const useDateStore = create<DateState>((set) => ({
           groupId,
           taskListId,
           getMonthDay(newDate),
+          userId,
         ],
       });
       return { pickDate: newDate };
